+++ src/socket.cc
@@ -222,7 +222,8 @@ void Socket::WarnUnlessImmediateOption(int32_t option) const {
 bool Socket::ValidateOpen() const {
     switch (state) {
     case State::Blocked:
-        ErrnoException(Env(), EBUSY).ThrowAsJavaScriptException();
+        ErrnoException(Env(), EBUSY, "Socket is blocked by a bind or unbind operation")
+            .ThrowAsJavaScriptException();
         return false;
     case State::Closed:
         ErrnoException(Env(), EBADF).ThrowAsJavaScriptException();
@@ -510,7 +511,10 @@ Napi::Value Socket::Send(const Napi::CallbackInfo& info) {
     if (!ValidateOpen()) return Env().Undefined();
 
     if (poller.Writing()) {
-        ErrnoException(Env(), EAGAIN).ThrowAsJavaScriptException();
+        ErrnoException(Env(), EBUSY,
+            "Socket is busy writing; only one send operation may be in progress "
+            "at any time")
+            .ThrowAsJavaScriptException();
         return Env().Undefined();
     }
 
@@ -542,7 +546,7 @@ Napi::Value Socket::Send(const Napi::CallbackInfo& info) {
            busy sending data to the I/O thread but is no longer able to respond
            to other events. */
 #ifdef ZMQ_NO_SYNC_RESOLVE
-        Warn(Env(), "Promise resolving by send() is delayed (ZMQ_NO_SYNC_RESOLVE).");
+        Warn(Env(), "Promise resolution by send() is delayed (ZMQ_NO_SYNC_RESOLVE).");
 #else
         if (sync_operations++ < max_sync_operations) {
             auto res = Napi::Promise::Deferred::New(Env());
@@ -575,7 +579,10 @@ Napi::Value Socket::Receive(const Napi::CallbackInfo& info) {
     if (!ValidateOpen()) return Env().Undefined();
 
     if (poller.Reading()) {
-        ErrnoException(Env(), EAGAIN).ThrowAsJavaScriptException();
+        ErrnoException(Env(), EBUSY,
+            "Socket is busy reading; only one receive operation may be in "
+            "progress at any time")
+            .ThrowAsJavaScriptException();
         return Env().Undefined();
     }
 
@@ -583,7 +590,7 @@ Napi::Value Socket::Receive(const Napi::CallbackInfo& info) {
         /* We can read from the socket immediately. This is a fast path.
            Also see the related comments in Send(). */
 #ifdef ZMQ_NO_SYNC_RESOLVE
-        Warn(Env(), "Promise resolving by receive() is delayed (ZMQ_NO_SYNC_RESOLVE).");
+        Warn(Env(), "Promise resolution by receive() is delayed (ZMQ_NO_SYNC_RESOLVE).");
 #else
         if (sync_operations++ < max_sync_operations) {
             auto res = Napi::Promise::Deferred::New(Env());