+++ test/unit/context-process-exit-test.ts
@@ -1,3 +1,4 @@
+import * as semver from "semver"
 import * as zmq from "../../src"
 
 import {assert} from "chai"
@@ -7,7 +8,7 @@ describe("context process exit", function() {
   describe("with default context", function() {
     it("should occur when sockets are closed", async function() {
       this.slow(200)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const socket1 = new zmq.Dealer()
         socket1.close()
         const socket2 = new zmq.Router()
@@ -19,7 +20,7 @@ describe("context process exit", function() {
 
     it("should occur when sockets are not closed", async function() {
       this.slow(200)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const socket1 = new zmq.Dealer()
         const socket2 = new zmq.Router()
       })
@@ -29,7 +30,7 @@ describe("context process exit", function() {
 
     it("should not occur when sockets are open and polling", async function() {
       this.slow(1000)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const socket1 = new zmq.Dealer()
         socket1.connect("inproc://foo")
         socket1.receive()
@@ -37,12 +38,61 @@ describe("context process exit", function() {
 
       assert.equal(code, -1)
     })
+
+    it("should produce warning when messages are queued with blocky", async function() {
+      this.slow(2000)
+      const {stderr} = await createProcess(() => {
+        zmq.context.blocky = true
+        const socket1 = new zmq.Dealer({linger: 1000})
+        socket1.connect("tcp://127.0.0.1:4567")
+        socket1.send(null)
+      })
+
+      if (semver.satisfies(zmq.version, ">= 4.2")) {
+        assert.match(
+          stderr.toString(),
+          /\(node:\d+\) WARNING: Waiting for queued ZeroMQ messages to be delivered\. Set 'context\.blocky = false' to change this behaviour\.\r?\n/,
+        )
+      } else {
+        assert.match(
+          stderr.toString(),
+          /\(node:\d+\) WARNING: Waiting for queued ZeroMQ messages to be delivered\.\r?\n/,
+        )
+      }
+    })
+
+    it("should produce warning when messages are queued without blocky", async function() {
+      this.slow(2000)
+      const {stderr} = await createProcess(() => {
+        zmq.context.blocky = false
+        const socket1 = new zmq.Dealer({linger: 1000})
+        socket1.connect("tcp://127.0.0.1:4567")
+        socket1.send(null)
+      })
+
+      assert.match(
+        stderr.toString(),
+        /\(node:\d+\) WARNING: Waiting for queued ZeroMQ messages to be delivered\.\r?\n/,
+      )
+    })
+
+    it("should not produce warning when messages are queued for a short time", async function() {
+      this.slow(1000)
+      const {stderr} = await createProcess(() => {
+        zmq.context.blocky = true
+        const socket1 = new zmq.Dealer({linger: 50})
+        socket1.connect("tcp://127.0.0.1:4567")
+        socket1.send(null)
+      })
+
+      assert.equal(stderr.toString(), "")
+    })
   })
 
   describe("with custom context", function() {
     it("should occur when sockets are closed", async function() {
       this.slow(200)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const context = new zmq.Context()
         const socket1 = new zmq.Dealer({context})
         socket1.close()
@@ -55,7 +105,7 @@ describe("context process exit", function() {
 
     it("should occur when sockets are closed and context is gced", async function() {
       this.slow(200)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         function run() {
           const context = new zmq.Context()
           const socket1 = new zmq.Dealer({context})
@@ -73,7 +123,7 @@ describe("context process exit", function() {
 
     it("should occur when sockets are not closed", async function() {
       this.slow(200)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const context = new zmq.Context()
         const socket1 = new zmq.Dealer({context})
         const socket2 = new zmq.Router({context})
@@ -84,7 +134,7 @@ describe("context process exit", function() {
 
     it("should not occur when sockets are open and polling", async function() {
       this.slow(1000)
-      const code = await createProcess(() => {
+      const {code} = await createProcess(() => {
         const context = new zmq.Context()
         const socket1 = new zmq.Dealer({context})
         socket1.connect("inproc://foo")