+++ test/unit/helpers.ts
@@ -14,7 +14,9 @@ if (semver.satisfies(zmq.version, ">= 4.2")) {
 /* Windows cannot bind on a ports just above 1014; start higher to be safe. */
 let seq = 5000
 
-export function uniqAddress(proto: string) {
+type Proto = "ipc" | "tcp" | "udp" | "inproc"
+
+export function uniqAddress(proto: Proto) {
   const id = seq++
   switch (proto) {
     case "ipc": {
@@ -31,7 +33,7 @@ export function uniqAddress(proto: string) {
   }
 }
 
-export function testProtos(...requested: string[]) {
+export function testProtos(...requested: Proto[]) {
   const set = new Set(requested)
 
   /* Do not test with ipc if unsupported. */
@@ -85,7 +87,13 @@ export async function createWorker<T, D extends {}>(
   })
 }
 
-export function createProcess(fn: () => void): Promise<number> {
+interface Result {
+  code: number
+  stdout: Buffer
+  stderr: Buffer
+}
+
+export function createProcess(fn: () => void): Promise<Result> {
   const src = `
     const zmq = require(${JSON.stringify(path.resolve(__dirname, "../.."))})
     const fn = ${fn.toString()}
@@ -96,20 +104,26 @@ export function createProcess(fn: () => void): Promise<number> {
   child.stdin.write(src)
   child.stdin.end()
 
-  child.stdout.on("data", (data: Buffer) => console.log(data.toString()))
-  child.stderr.on("data", (data: Buffer) => console.error(data.toString()))
+  let stdout: Buffer = Buffer.alloc(0)
+  let stderr: Buffer = Buffer.alloc(0)
+  child.stdout.on("data", (data: Buffer) => {
+    stdout = Buffer.concat([stdout, data])
+  })
+  child.stderr.on("data", (data: Buffer) => {
+    stderr = Buffer.concat([stderr, data])
+  })
 
   return new Promise((resolve, reject) => {
     child.on("close", (code: number, signal: string) => {
       if (signal != null) {
         reject(new Error(`Child exited with ${signal}`))
       } else {
-        resolve(code)
+        resolve({code, stdout, stderr})
       }
     })
 
     setTimeout(() => {
-      resolve(-1)
+      resolve({code: -1, stdout, stderr})
       child.kill()
     }, 750)
   })